bootstrap moved to modules subdirectory (#12)
